+++ ui/apps/platform/src/Containers/ComplianceEnhanced/Status/Reports/ComplianceReportsClusterPage.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import ComplianceReports from './ComplianceReports';
+
+function ComplianceReportsClusterPage() {
+    return (
+        <>
+            <ComplianceReports type="cluster" />
+        </>
+    );
+}
+
+export default ComplianceReportsClusterPage;