+++ ui/apps/platform/src/Containers/ComplianceEnhanced/Status/Reports/ComplianceReportsScanPage.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import ComplianceReports from './ComplianceReports';
+
+function ComplianceReportsScanPage() {
+    return (
+        <>
+            <ComplianceReports type="scan" />
+        </>
+    );
+}
+
+export default ComplianceReportsScanPage;