+++ ui/apps/platform/src/Containers/MainPage/Body.tsx
@@ -91,7 +91,7 @@ const routeComponentMap: Record<RouteKey, RouteComponent> = {
     },
     'compliance-enhanced': {
         component: asyncComponent(
-            () => import('Containers/ComplianceEnhanced/Dashboard/ComplianceDashboardPage')
+            () => import('Containers/ComplianceEnhanced/Status/ComplianceStatusPage')
         ),
         path: complianceEnhancedBasePath,
     },